% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R, R/zzz.R
\docType{package}
\name{windowlickr}
\alias{windowlickr}
\title{windowlickr}
\usage{
windowlickr(
  bcf,
  func,
  windowsize = NULL,
  slide = windowsize,
  samples = NULL,
  minMAF = 0.1,
  maxMissing = 0.8,
  windows = NULL,
  chrom = NULL,
  from = NULL,
  to = NULL,
  errors = "stop",
  export = NULL,
  pkgs = NULL
)
}
\arguments{
\item{bcf}{Indexed BCF or VCF file}

\item{func}{Function to evaulate for each window. is called as func(window.metadata, window.genotypes) for each window. Func ideally returns a tibble or similar.}

\item{windowsize}{Size of each window (required unless `windows` provided)}

\item{slide}{Number of bases to skip between each window start (default `windowsize`)}

\item{samples}{Set of samples to calculate LD among}

\item{minMAF}{Minimum SNP minor allle freq}

\item{maxMissing}{Maximum SNP missing data rate}

\item{windows}{Restrict analyses to these windows (expected to be created with bcf_getWindows)}

\item{chrom}{Only generate windows on contig 'chrom'}

\item{from}{Only generate windows starting at `from` on `chrom`}

\item{to}{Only generate windows until `to` on `chrom`}

\item{errors}{One of "remove" or "stop", see ?foreach::foreach 's .errorhandling parameter}

\item{export}{Vector of (string) variable names to export to the foreach loop}

\item{pkgs}{Packages required to within `func`, require()'d by foreach loop}
}
\value{
list One list per window, with slots `regions`, `contig`, `start`, `stop`, `result`, where result is the return value of `func` applied to this window.
}
\description{
Windowed analyses of BCF files.
}
\author{
Kevin Murray
}
